Fix check patch and fs tests

+++ file-system/index.js
@@ -4,41 +4,57 @@ import { toArray } from '../utils/index.js'
 
 export function fileSystem() {
   return {
-    async delete(files) {
-      files = toArray(files)
+    async delete(path) {
+      if (Array.isArray(path)) {
+        let paths = toArray(path)
 
-      return await Promise.all(
-        files.map(async (path) => {
-          return await fs.unlink(path)
-        })
-      )
+        return await Promise.all(
+          paths.map(async (path) => {
+            return await fs.unlink(path)
+          })
+        )
+      }
+
+      return await fs.unlink(path)
     },
 
-    async read(files) {
-      files = toArray(files)
+    async read(path) {
+      if (Array.isArray(path)) {
+        let paths = toArray(path)
 
-      let promises = await Promise.all(
-        files.map(async (path) => {
-          try {
-            let content = await fs.readFile(path)
-            return { path, content }
-          } catch (err) {
-            return null
-          }
-        })
-      )
+        return (
+          await Promise.all(
+            paths.map(async (path) => {
+              try {
+                let content = await fs.readFile(path)
+                return { path, content }
+              } catch (err) {
+                return null
+              }
+            })
+          )
+        ).filter(Boolean)
+      }
 
-      return promises.filter(Boolean)
+      try {
+        return await fs.readFile(path)
+      } catch (err) {
+        return null
+      }
     },
 
-    async write(files) {
-      files = toArray(files)
+    async write(path, content) {
+      if (Array.isArray(path)) {
+        let paths = toArray(path)
 
-      return await Promise.all(
-        toArray(files).map(async ({ path, content }) => {
-          return await fs.writeFile(path, content)
-        })
-      )
+        return await Promise.all(
+          paths.map(async ({ path, content }) => {
+            return await fs.writeFile(path, content)
+          })
+        )
+      }
+
+      return await fs.writeFile(path, content)
     },
   }
 }
